Better Contentful query experience
