+++ src/app/[locale]/(subpages)/blog/page.tsx
@@ -2,14 +2,14 @@ import { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { PageProps } from "~/@types";
 import { BlogPaginator, BlogPostPreview } from "~/components";
-import { getPaginatedPostsFromCache } from "~/utils";
+import { query } from "~/utils";
 
 export const metadata: Metadata = {
   title: "Blog",
 };
 
 async function getData({ params: { locale }, searchParams: { page, size } }: PageProps) {
-  const { items, currentPage, pageSize, totalItems, totalPages } = await getPaginatedPostsFromCache(page, size, locale);
+  const { items, currentPage, pageSize, totalItems, totalPages } = await query.paginatedPosts(locale, page, size);
 
   return {
     posts: items.filter((post) => !post.hidden),
