+++ src/utils/contentful/index.ts
@@ -1,4 +1,3 @@
 export * from "./contentful-client";
-export * from "./contentful-get";
 export * from "./contentful-renderer";
-export * from "./contentful-functions";
+export * from "./contentful-query";
